Fix bug related to limit feature
This commit fixes a bug related to the interaction between the limit feature and the rest of features. In the listTodos feature, limit was the first query param that was checked. This worked on its own, but when used in conjunction with other parameters, it did not work correctly. For example, if a user requested a list of todos of owner "Fry" and limit 5, what would happen is listTodos would first limit all of the todos to 10 and then grab the ones with Fry out of those 10. This is opposed to what is supposed happen, as what is supposed to happen is that it should first filter by owner, and then grab 10 of of the todos.
Co-authored-by: jadynsondrol <sondr037@morris.umn.edu>